requirements.txt
requests
bs4
pandas
numpy
pyarrow
google-cloud-bigquery
db-dtypes
pip install openai==0.28